BUG: suppress all uses of ANSI terminal escapes when not supported
